+++ tests/test_sign_cmd.py
@@ -1,8 +1,8 @@
 import pytest
 
-from application_client.boilerplate_transaction import Transaction
-from application_client.boilerplate_command_sender import BoilerplateCommandSender, Errors
-from application_client.boilerplate_response_unpacker import unpack_get_public_key_response, unpack_sign_tx_response
+from application_client.klaytn_transaction import Transaction
+from application_client.klaytn_command_sender import KlaytnCommandSender, Errors
+from application_client.klaytn_response_unpacker import unpack_get_public_key_response, unpack_sign_tx_response
 from ragger.error import ExceptionRAPDU
 from ragger.navigator import NavInsID
 from utils import ROOT_SCREENSHOT_PATH, check_signature_validity
@@ -15,7 +15,7 @@ from utils import ROOT_SCREENSHOT_PATH, check_signature_validity
 # We will ensure that the displayed information is correct by using screenshots comparison
 def test_sign_tx_short_tx(firmware, backend, navigator, test_name):
     # Use the app interface instead of raw interface
-    client = BoilerplateCommandSender(backend)
+    client = KlaytnCommandSender(backend)
     # The path used for this entire test
     path: str = "m/44'/1'/0'/0/0"
 
@@ -61,7 +61,7 @@ def test_sign_tx_short_tx(firmware, backend, navigator, test_name):
 # In particular the long memo will force the transaction to be sent in multiple chunks
 def test_sign_tx_long_tx(firmware, backend, navigator, test_name):
     # Use the app interface instead of raw interface
-    client = BoilerplateCommandSender(backend)
+    client = KlaytnCommandSender(backend)
     path: str = "m/44'/1'/0'/0/0"
 
     rapdu = client.get_public_key(path=path)
@@ -100,7 +100,7 @@ def test_sign_tx_long_tx(firmware, backend, navigator, test_name):
 # The test will ask for a transaction signature that will be refused on screen
 def test_sign_tx_refused(firmware, backend, navigator, test_name):
     # Use the app interface instead of raw interface
-    client = BoilerplateCommandSender(backend)
+    client = KlaytnCommandSender(backend)
     path: str = "m/44'/1'/0'/0/0"
 
     rapdu = client.get_public_key(path=path)